scripts/requirements.txt
anthropic==0.94.0
numpy==2.4.4
pdfplumber==0.11.9
python-dotenv==1.2.2
rank-bm25==0.2.2
sentence-transformers==5.4.0
